+++ carl/envs/brax/carl_double_pendulum.py
@@ -2,8 +2,8 @@ from typing import Any, Dict, List, Optional, Union
 
 import numpy as np
 import jax.numpy as jnp
-from brax.envs.half_cheetah import Halfcheetah
 from brax.envs import create
+from brax.envs.inverted_double_pendulum import InvertedDoublePendulum
 from carl.envs.brax.brax_wrappers import GymWrapper, VectorGymWrapper
 
 from carl.context.selection import AbstractSelector
@@ -14,36 +14,44 @@ from carl.envs.brax.carl_brax_env import CARLBraxEnv
 
 DEFAULT_CONTEXT = {
     "stiffness_factor": 1,
-    "gravity": -9.81,
-    "friction": 0.4,
+    "gravity_x": 1e-5,
+    "gravity_z": -9.81,
+    "friction": 0.8,
     "damping_factor": 1,
     "actuator_strength_factor": 1,
-    "torso_mass": 6.25,
+    "mass_cart": 10.5,
+    "mass_pole_0": 4.2,
+    "mass_pole_1": 4.2,
     "dt": 0.01
 }
 
 CONTEXT_BOUNDS = {
     "stiffness_factor": (0, np.inf, float),
-    "gravity": (-np.inf, -0.1, float),
+    "gravity_x": (-np.inf, np.inf, float),
+    "gravity_z": (-np.inf, -0.1, float),
     "friction": (-np.inf, np.inf, float),
     "damping_factor": (-np.inf, np.inf, float),
     "actuator_strength_factor": (1, np.inf, float),
-    "torso_mass": (0.1, np.inf, float),
+    "mass_cart": (0.1, np.inf, float),
+    "mass_pole_0": (0.1, np.inf, float),
+    "mass_pole_1": (0.1, np.inf, float),
     "dt": (0.0001, 0.03, float),
 }
 
 
 
-class CARLHalfcheetah(CARLBraxEnv):
-    env_name: str = "halfcheetah"
+class CARLInvertedDoublePendulum(CARLBraxEnv):
+    env_name: str = "inverted_double_pendulum"
     DEFAULT_CONTEXT: Context = DEFAULT_CONTEXT
-
+    
     def _update_context(self) -> None:
-        self.env: Halfcheetah
+        self.env: InvertedDoublePendulum
         config = {}
-        config["gravity"] = jnp.array([0, 0, self.context["gravity"]])
+        config["gravity"] = jnp.array([self.context["gravity_x"], 0, self.context["gravity_z"]])
         config["dt"] = jnp.array(self.context["dt"])
-        new_mass = self.env._env.sys.link.inertia.mass.at[0].set(self.context["torso_mass"])
+        new_mass = self.env._env.sys.link.inertia.mass.at[0].set(self.context["mass_cart"])
+        new_mass = new_mass.at[1].set(self.context["mass_pole_0"])
+        new_mass = new_mass.at[2].set(self.context["mass_pole_1"])
         # TODO: do we want to implement this?
         #new_com = self.env.sys.link.inertia.transform
         #new_inertia = self.env.sys.link.inertia.i
@@ -58,4 +66,3 @@ class CARLHalfcheetah(CARLBraxEnv):
         geoms[0] = geoms[0].replace(friction=jnp.array([self.context["friction"]]))
         config["geoms"] = geoms
         self.env._env.sys = self.env._env.sys.replace(**config)
-