+++ carl/envs/carl_env.py
@@ -2,26 +2,22 @@ from __future__ import annotations
 
 from typing import Any, Dict, List, Mapping, Optional, Tuple, Type, Union
 
-import importlib
 import inspect
 import json
 import os
 from types import ModuleType
 
-import gym
+# TODO: this is only needed for brax,remove
+import gym as legacy_gym
+import gymnasium as gym
 import numpy as np
-from gym import Wrapper, spaces
+from gymnasium import Wrapper, spaces
 
 from carl.context.augmentation import add_gaussian_noise
 from carl.context.selection import AbstractSelector, RoundRobinSelector
 from carl.context.utils import get_context_bounds
 from carl.utils.trial_logger import TrialLogger
 from carl.utils.types import Context, Contexts, ObsType, Vector
-
-brax_spec = importlib.util.find_spec("brax")
-if brax_spec is not None:
-    import jax.numpy as jnp
-    import jaxlib
 
 
 class CARLEnv(Wrapper):
@@ -86,6 +82,7 @@ class CARLEnv(Wrapper):
 
     available_scale_methods = ["by_default", "by_mean", "no"]
     available_instance_modes = ["random", "rr", "roundrobin"]
+    metadata = {"render_modes": ["human", "rgb_array"]}
 
     def __init__(
         self,
@@ -111,6 +108,7 @@ class CARLEnv(Wrapper):
         # Gather args
         self._context: Context  # init for property
         self._contexts: Contexts  # init for property
+
         self.default_context = default_context
         self.contexts = contexts
         self.context_mask = context_mask
@@ -254,8 +252,14 @@ class CARLEnv(Wrapper):
         self._contexts = {
             k: self.fill_context_with_default(context=v) for k, v in contexts.items()
         }
+        return
 
-    def reset(self, **kwargs: Dict) -> Union[ObsType, tuple[ObsType, dict]]:  # type: ignore [override]
+    def reset(
+        self,
+        seed: int | None = None,
+        options: dict[str, Any] | None = None,
+        **kwargs: Dict,
+    ) -> Union[ObsType, tuple[ObsType, dict]]:  # type: ignore [override]
         """
         Reset environment.
 
@@ -280,27 +284,26 @@ class CARLEnv(Wrapper):
             self._progress_instance()
         self._update_context()
         self._log_context()
-        return_info = kwargs.get("return_info", False)
-        _ret = self.env.reset(**kwargs)  # type: ignore [arg-type]
+        _ret = self.env.reset(seed=seed, options=options, **kwargs)  # type: ignore [arg-type]
         info_dict = dict()
+
         if return_info:
             state, info_dict = _ret
             info_dict["context_key"] = self.context_key
         else:
             state = _ret
         state = self.build_context_adaptive_state(state=state)
-        ret = state
-        if return_info:
-            ret = state, info_dict
-        return ret
+        return state, info_dict
 
     def build_context_adaptive_state(
         self, state: List[float], context_feature_values: Optional[Vector] = None
     ) -> Union[Vector, Dict]:
         tnp: ModuleType = np
+          
         if brax_spec is not None:
             if type(state) == jaxlib.xla_extension.ArrayImpl:
                 tnp = jnp
+          
         if not self.hide_context:
             if context_feature_values is None:
                 # use current context
@@ -328,7 +331,7 @@ class CARLEnv(Wrapper):
                 state = tnp.concatenate((state, context_values))
         return state
 
-    def step(self, action: Any) -> Tuple[Any, Any, bool, Dict]:
+    def step(self, action: Any) -> Tuple[Any, Any, bool, bool, Dict]:
         """
         Step the environment.
 
@@ -381,6 +384,7 @@ class CARLEnv(Wrapper):
             done = True
         info["context_key"] = self.context_key
         return state, reward, done, info
+
 
     def __getattr__(self, name: str) -> Any:
         # TODO: does this work with activated noise? I think we need to update it
@@ -484,7 +488,10 @@ class CARLEnv(Wrapper):
         self.observation_space: gym.spaces.Space
         if (
             not self.dict_observation_space
-            and not isinstance(self.observation_space, spaces.Box)
+            and not (
+                isinstance(self.observation_space, spaces.Box)
+                or isinstance(self.observation_space, legacy_gym.spaces.Box)
+            )
             and not self.hide_context
         ):
             raise ValueError(