+++ carl/envs/classic_control/carl_acrobot.py
@@ -1,7 +1,7 @@
 from typing import Dict, List, Optional, Union
 
 import numpy as np
-from gym.envs.classic_control import AcrobotEnv
+from gymnasium.envs.classic_control import AcrobotEnv
 
 from carl.context.selection import AbstractSelector
 from carl.envs.carl_env import CARLEnv
@@ -73,7 +73,6 @@ class CustomAcrobotEnv(AcrobotEnv):
         self,
         *,
         seed: Optional[int] = None,
-        return_info: bool = False,
         options: Optional[dict] = None,
     ) -> Union[np.ndarray, tuple[np.ndarray, dict]]:
         super().reset(seed=seed)
@@ -90,10 +89,7 @@ class CustomAcrobotEnv(AcrobotEnv):
             self.INITIAL_VELOCITY_UPPER,
         )
         self.state = self.np_random.uniform(low=low, high=high).astype(np.float32)
-        if not return_info:
-            return self._get_ob()
-        else:
-            return self._get_ob(), {}
+        return self._get_ob(), {}
 
 
 class CARLAcrobotEnv(CARLEnv):