+++ carl/envs/rna/rna_environment.py
@@ -12,8 +12,7 @@ from distance import hamming
 import numpy as np
 
 from RNA import fold
-import gym
-
+import gymnasium as gym
 from typing import Any, List
 
 
@@ -97,7 +96,6 @@ def _encode_dot_bracket(  # type: ignore[no-untyped-def]
 
 
 def _encode_pairing(secondary: str):  # type: ignore[no-untyped-def]
-
     pairing_encoding = [None] * len(secondary)
     stack = []
     for index, symbol in enumerate(secondary, 0):