+++ changelog.md
@@ -1,4 +1,5 @@
 # 0.2.1
+- Make sampling of contexts deterministic with seed
 - Add Finger (DMC) env
 - Readd RNA env (#78)